move firebase scripts
