+++ process-dataset/firebase/setup-firebase.js
